+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Cagnardi 2018 cefazolin ([doi:10.3389/fphar.2018.01137](https://doi.org/10.3389/fphar.2018.01137)) -- client-owned dogs given surgical antimicrobial prophylaxis.
+
 - Add Palmer 2025 moxifloxacin ([doi:10.1002/bcp.70005](https://doi.org/10.1002/bcp.70005)) -- children with rifampicin-resistant tuberculosis.
 
 - Add Nguyen 2026 dupilumab ([doi:10.1002/cpt.70233](https://doi.org/10.1002/cpt.70233)) -- healthy adults and adults, adolescents and children aged 1 year and older with eosinophilic esophagitis.